Begining to work on validation

+++ Transactions/Controllers/TransactionsController.cs
@@ -1,7 +1,9 @@
 using System;
 using System.Collections.Generic;
+using System.ComponentModel;
 using System.IO;
 using System.Linq;
+using System.Reflection;
 using System.Text;
 using System.Threading.Tasks;
 using System.Transactions;
@@ -15,6 +17,7 @@ using Transactions.Database.Entities;
 using Transactions.Mappings;
 using Transactions.Mappings.Entities;
 using Transactions.Models.Transaction.Enums;
+using Transactions.Problems;
 using Transactions.Services;
 
 namespace Transactions.Controllers{
@@ -48,19 +51,62 @@ namespace Transactions.Controllers{
             CsvParserOptions csvParserOptions = new CsvParserOptions(true, ',');
             CsvTransactionMapping csvTransactionMapping = new CsvTransactionMapping();
             CsvParser<TransactionCsvEntity> csvParser = new CsvParser<TransactionCsvEntity>(csvParserOptions,csvTransactionMapping);
-    
+
             string filePath = await GetFilePath(csvFile);
 
             var transactionList = csvParser.ReadFromFile(filePath, Encoding.ASCII).ToList();
 
+            var validationProblem = Validate(transactionList);
+
+            if(validationProblem.Errors.Count>0){
+                return BadRequest(validationProblem);
+            }
+
             var transactions = _mapper.Map<List<CsvMappingResult<TransactionCsvEntity>>, List<Models.Transaction.Transaction>>(transactionList);
 
             transactions.RemoveAt(transactions.Count-1);
-            await _transactionsService.InsertTransactions(transactions);
+            //await _transactionsService.InsertTransactions(transactions);
 
             System.IO.File.Delete(filePath);
 
             return Ok("Transaction splitted");
+        }
+
+        private ValidationProblem Validate(List<CsvMappingResult<TransactionCsvEntity>> list){
+            List<Errors> errors = new List<Errors>();
+
+            foreach(var item in list){
+                if(string.IsNullOrEmpty(item.Result.Id)){
+                    errors.Add(CreateError("id", ErrEnum.Required, GetEnumDescription(ErrEnum.Required)));
+                    continue;
+                }
+            }
+
+            return new ValidationProblem{
+                Errors = errors
+            };
+        }
+
+        private Errors CreateError(string tag, ErrEnum err, string message){
+            return new Errors{
+                Tag = tag,
+                Error = err,
+                Message = message
+            };
+        }
+
+        private string GetEnumDescription(Enum value)
+        {
+            FieldInfo fi = value.GetType().GetField(value.ToString());
+
+            DescriptionAttribute[] attributes = fi.GetCustomAttributes(typeof(DescriptionAttribute), false) as DescriptionAttribute[];
+
+            if (attributes != null && attributes.Any())
+            {
+                return attributes.First().Description;
+            }
+
+            return value.ToString();
         }
 
         [HttpGet]
